+++ src/main/java/heavysnow/heath/controller/PostController.java
@@ -7,8 +7,8 @@ import heavysnow.heath.dto.comment.CommentUpdateDto;
 import heavysnow.heath.dto.post.PostAddRequest;
 import heavysnow.heath.dto.post.PostAddResponse;
 import heavysnow.heath.dto.post.PostEditRequest;
-import heavysnow.heath.dto.postdto.PostDetailResponse;
-import heavysnow.heath.dto.postdto.PostListResponse;
+import heavysnow.heath.dto.post.PostDetailResponse;
+import heavysnow.heath.dto.post.PostListResponse;
 import heavysnow.heath.exception.UnauthorizedException;
 import heavysnow.heath.service.CommentService;
 import heavysnow.heath.service.LikedService;
