+++ src/main/java/heavysnow/heath/dto/post/PostListResponse.java
@@ -1,4 +1,4 @@
-package heavysnow.heath.dto.postdto;
+package heavysnow.heath.dto.post;
 
 import heavysnow.heath.domain.Post;
 import lombok.AccessLevel;
