+++ src/main/java/heavysnow/heath/service/PostService.java
@@ -8,8 +8,8 @@ import heavysnow.heath.dto.PostDatesResponseDto;
 import heavysnow.heath.dto.post.PostAddRequest;
 import heavysnow.heath.dto.post.PostAddResponse;
 import heavysnow.heath.dto.post.PostEditRequest;
-import heavysnow.heath.dto.postdto.PostDetailResponse;
-import heavysnow.heath.dto.postdto.PostListResponse;
+import heavysnow.heath.dto.post.PostDetailResponse;
+import heavysnow.heath.dto.post.PostListResponse;
 import heavysnow.heath.exception.ForbiddenException;
 import heavysnow.heath.exception.NotFoundException;
 import heavysnow.heath.repository.MemberRepository;
