+++ src/test/java/heavysnow/heath/service/PostServiceTest.java
@@ -4,8 +4,8 @@ import heavysnow.heath.dto.comment.CommentCreateDto;
 import heavysnow.heath.dto.MemberDto;
 import heavysnow.heath.dto.post.PostAddRequest;
 import heavysnow.heath.dto.post.PostEditRequest;
-import heavysnow.heath.dto.postdto.PostDetailResponse;
-import heavysnow.heath.dto.postdto.PostListResponse;
+import heavysnow.heath.dto.post.PostDetailResponse;
+import heavysnow.heath.dto.post.PostListResponse;
 import heavysnow.heath.exception.NotFoundException;
 import heavysnow.heath.repository.CommentRepository;
 import heavysnow.heath.repository.MemberPostLikedRepository;
